LayoutTests/http/tests/ssl/applepay/ApplePayMerchantValidationEvent.https.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<script src="/js-test-resources/ui-helper.js"></script>
<script src="/resources/js-test-pre.js"></script>
<script src="/resources/payment-request.js"></script>
</head>
<body>
<script>

description("Test the applepayvalidatemerchant event.");

window.jsTestIsAsync = true;

function validPaymentMethod() {
    return {
        supportedMethods: 'https://apple.com/apple-pay',
        data: {
            version: 2,
            merchantIdentifier: '',
            countryCode: 'US',
            supportedNetworks: ['visa', 'masterCard'],
            merchantCapabilities: ['supports3DS'],
        },
    }
}

function validPaymentDetails() {
    return {
        total: { label: 'Your label', amount: { currency: 'USD', value: '10.00' } },
    }
}

function runTest(test) {
    return new Promise((resolve, reject) => {
        var paymentRequest = new PaymentRequest([validPaymentMethod()], validPaymentDetails());
        paymentRequest.onapplepayvalidatemerchant = (event) => {
            test(event, paymentRequest);
            paymentRequest.abort().catch(() => {});
            resolve();
        };
        activateThen(() => paymentRequest.show().catch(() => {}));
    });
}

async function go() {
    debug("Testing ApplePayMerchantValidationEvent")
    debug("")

    await runTest(event => shouldThrow(() => event.complete(undefined)));
    await runTest(event => shouldThrow(() => event.complete(null)));
    await runTest(event => shouldThrow(() => event.complete([])));
    await runTest(event => shouldThrow(() => event.complete('')));
    await runTest(event => shouldThrow(() => event.complete(7)));
    await runTest(event => shouldThrow(() => { event.complete({}); event.complete({}); }));
    await runTest((event, paymentRequest) => shouldThrow(() => { paymentRequest.abort(); event.complete({}); }));
    await runTest(event => shouldNotThrow(() => event.complete({})));
    await runTest(event => shouldBe("event.validationURL", "'https://webkit.org/'"));

    document.querySelector("button").remove();
    finishJSTest();
}

window.onload = function() {
    var button = document.querySelector("button");
    let x = button.offsetLeft + 2;
    let y = button.offsetTop + 2;
    UIHelper.activateAt(x, y);
}

</script>
<button onclick='go()'>Go</button>
<script src="/resources/js-test-post.js"></script>
</body>
</html>
