LayoutTests/http/tests/ssl/applepay/ApplePayMerchantValidationEvent.https-expected.txt

Test the applepayvalidatemerchant event.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Testing ApplePayMerchantValidationEvent

PASS () => event.complete(undefined) threw exception TypeError: Type error.
PASS () => event.complete(null) threw exception TypeError: Type error.
PASS () => event.complete([]) threw exception TypeError: Type error.
PASS () => event.complete('') threw exception TypeError: Type error.
PASS () => event.complete(7) threw exception TypeError: Type error.
PASS () => { event.complete({}); event.complete({}); } threw exception InvalidStateError: The object is in an invalid state..
PASS () => { paymentRequest.abort(); event.complete({}); } threw exception InvalidStateError: The object is in an invalid state..
PASS () => event.complete({}) did not throw exception.
PASS event.validationURL is 'https://webkit.org/'
PASS successfullyParsed is true

TEST COMPLETE
